+++ packages/sampler/stories/moonstone-stories/VideoPlayer.js
@@ -123,6 +123,7 @@ storiesOf('Moonstone', module)
 						muted={boolean('muted', Config, true)}
 						noAutoPlay={boolean('noAutoPlay', Config)}
 						noAutoShowMediaControls={boolean('noAutoShowMediaControls', Config)}
+						noMediaSliderFeedback={boolean('noMediaSliderFeedback', Config, false)}
 						noMiniFeedback={boolean('noMiniFeedback', Config)}
 						noSlider={boolean('noSlider', Config)}
 						pauseAtEnd={boolean('pauseAtEnd', Config)}