+++ packages/moonstone/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `moonstone/VideoPlayer` property `noMediaSliderFeedback`
+
 ### Fixed
 
 - `moonstone/VideoPlayer` to round the time displayed down to nearest second